Initiate exceeds requirements

+++ routes/course/delete-course.js
@@ -1,6 +1,16 @@
 const deleteCourse = (router, authenticateUser, asyncHandler, Course) => {
     router.delete('/courses/:id', authenticateUser, asyncHandler(async (req, res) => {
-        const course = await Course.destroy({
+        const course = await Course.findOne({
+            where: {
+                id: req.params.id
+            }
+        });
+        if (req.currentUser.id !== course.userId) {
+            const message = 'Access Denied: Only course owner may delete course';
+            console.error(message);
+            return res.status(403).json({ message });
+        }
+        await Course.destroy({
             where: { 
                 id: req.params.id
             }
